+++ packages/dart/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `normalizeFile` method
 - Add `headFile` method
 - Add `tailFile` method
+- Add `removeLocalePrefix` method
 
 ## 1.1.8 (2025-11-13)